Revert "use only primitive types in packet definitions"
This reverts commit 3df318d66b0fbdb5b1524549b107baa83131f50e.

+++ src/Core/Extensions/ConnectionExtensions.cs
@@ -11,7 +11,7 @@ public static class ConnectionExtensions
         connection.Phase = phase;
         connection.Send(new GCPhase
         {
-            Phase = (byte)connection.Phase
+            Phase = connection.Phase
         });
     }
 }
